requirements.txt
setuptools
django
django-rest-framework
pillow
ipython
isort
